# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AfterSale Copilot — Local Simulation",
    "description": "Stages 1–3: single Agent with deterministic after-sales controls. No real payments. Business APIs require authenticated sessions and CSRF. Amounts are CNY strings.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/actions": {
      "post": {
        "tags": [
          "stage2"
        ],
        "summary": "Submit",
        "operationId": "submit_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          }
        }
      }
    },
...
    "/portal/actions": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "Actions",
        "operationId": "actions_portal_actions_get",
        "parameters": [
          {
            "name": "risk",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RiskLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExecutionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionRead"
                  },
                  "title": "Response Actions Portal Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/portal/runs": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "Runs",
        "operationId": "runs_portal_runs_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunRead"
                  },
                  "title": "Response Runs Portal Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionCreate": {
        "properties": {
          "ticket_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Ticket Id"
          },
          "proposed_action": {
            "$ref": "#/components/schemas/ActionType"
          },
          "order_item_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Item Id"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "evidence_provided": {
            "type": "boolean",
            "title": "Evidence Provided",
            "default": false
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Idempotency Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ticket_id",
          "proposed_action",
          "idempotency_key"
        ],
        "title": "ActionCreate"
      },
      "ActionRead": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/PolicyDecision"
          },
          "authorized_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "type": "string",
            "title": "Amount",
            "default": "0.00"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel",
            "default": "LOW"
          },
          "rule_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rule Codes"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "order_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Item Id"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "proposed_action": {
            "$ref": "#/components/schemas/ActionType"
          },
          "final_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution_status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "evidence_provided": {
            "type": "boolean",
            "title": "Evidence Provided"
          },
          "return_received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Received At"
          },
          "snapshot_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "rule_codes",
          "explanation",
          "policy_version",
          "id",
          "ticket_id",
          "user_id",
          "order_id",
          "order_item_id",
          "quantity",
          "idempotency_key",
          "proposed_action",
          "final_action",
          "execution_status",
          "evidence_provided",
          "return_received_at",
          "snapshot_json",
          "created_at",
          "updated_at"
        ],
        "title": "ActionRead"
      },
      "ActionType": {
        "type": "string",
        "enum": [
          "CANCEL_AND_REFUND",
          "RETURN_AND_REFUND",
          "REFUND_MISSING_ITEM",
          "REPLACE_ITEM",
          "CREATE_LOGISTICS_TICKET"
        ],
        "title": "ActionType"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorRead": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorRead"
      },
...
      "ExecutionStatus": {
        "type": "string",
        "enum": [
          "BLOCKED",
          "WAITING_APPROVAL",
          "WAITING_RETURN",
          "READY",
          "SUCCESS",
          "FAILED",
          "REJECTED"
        ],
        "title": "ExecutionStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyDecision": {
        "type": "string",
        "enum": [
          "ALLOW",
          "DENY",
          "REQUIRE_APPROVAL",
          "NEED_MORE_INFO"
        ],
        "title": "PolicyDecision"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "RiskLevel"
      },
      "RunRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "ticket_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Id"
          },
          "user_query": {
            "type": "string",
            "title": "User Query"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "llm_calls": {
            "type": "integer",
            "title": "Llm Calls"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "user_id",
          "order_id",
          "ticket_id",
          "user_query",
          "reply",
          "outcome",
          "status",
          "llm_calls",
          "latency_ms",
          "created_at",
          "parent_run_id"
        ],
        "title": "RunRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}